+++ apps/app/src/components/token/token-portfolio.tsx
@@ -15,14 +15,10 @@
 
 "use client";
 
-import { getPortfolio } from "@lightdotso/client";
 import { Number } from "@lightdotso/ui";
-import { useSuspenseQuery, useQueryClient } from "@tanstack/react-query";
 import type { FC } from "react";
 import type { Address } from "viem";
-import type { TokenPortfolioData } from "@/data";
-import { queryKeys } from "@/queryKeys";
-import { useAuth } from "@/stores";
+import { useSuspenseQueryPortfolio } from "@/query";
 
 // -----------------------------------------------------------------------------
 // Props
@@ -44,58 +40,18 @@ export const TokenPortfolio: FC<TokenPortfolioProps> = ({
   isNeutral = false,
 }) => {
   // ---------------------------------------------------------------------------
-  // Stores
-  // ---------------------------------------------------------------------------
-
-  const { clientType } = useAuth();
-
-  // ---------------------------------------------------------------------------
   // Query
   // ---------------------------------------------------------------------------
 
-  const queryClient = useQueryClient();
-
-  const currentData: TokenPortfolioData | undefined = queryClient.getQueryData(
-    queryKeys.portfolio.get({ address }).queryKey,
-  );
-
-  const { data: portfolio } = useSuspenseQuery<TokenPortfolioData | null>({
-    queryKey: queryKeys.portfolio.get({ address }).queryKey,
-    queryFn: async () => {
-      if (!address) {
-        return null;
-      }
-
-      const res = await getPortfolio(
-        {
-          params: {
-            query: {
-              address: address,
-            },
-          },
-        },
-        clientType,
-      );
-
-      // Return if the response is 200
-      return res.match(
-        data => {
-          return data;
-        },
-        _ => {
-          return currentData ?? null;
-        },
-      );
-    },
-  });
-
-  if (!portfolio) {
-    return null;
-  }
+  const { portfolio } = useSuspenseQueryPortfolio({ address: address });
 
   // ---------------------------------------------------------------------------
   // Render
   // ---------------------------------------------------------------------------
+
+  if (!portfolio) {
+    return null;
+  }
 
   return (
     portfolio.balances && (
